fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.10
python-multipart==0.0.12
Pillow==11.0.0
insightface==0.7.3
easyocr==1.7.1
numpy==1.26.4
opencv-python==4.10.0.84
onnxruntime==1.19.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.14.0
boto3==1.35.0
